Add .editorconfig, move jsonata.js to src
