+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jsonata",
-  "version": "1.5.4",
+  "version": "1.5.5",
   "description": "JSON query and transformation language",
   "module": "jsonata.js",
   "main": "jsonata-es5.js",
@@ -14,14 +14,15 @@
     "pretest": "npm run lint",
     "mocha": "node ./node_modules/istanbul/lib/cli.js cover --report cobertura --report html ./node_modules/mocha/bin/_mocha -- \"test/**/*.js\"",
     "test": "npm run mocha",
-    "posttest": "npm run check-coverage && npm run minify && npm run babel && npm run browserify && npm run minify-es5",
+    "posttest": "npm run check-coverage && npm run copy && npm run minify && npm run babel && npm run browserify && npm run minify-es5",
+    "copy": "ncp src/jsonata.js ./jsonata.js",
     "check-coverage": "istanbul check-coverage -statement 100 -branch 100 -function 100 -line 100",
     "minify": "uglifyjs jsonata.js -o jsonata.min.js --compress --mangle",
-    "lint": "eslint .",
+    "lint": "eslint ./src",
     "doc": "jsdoc --configure jsdoc.json .",
     "cover": "istanbul cover _mocha",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
-    "babel": "node ./node_modules/babel-cli/bin/babel.js --out-file jsonata-es5.js jsonata.js",
+    "babel": "node ./node_modules/babel-cli/bin/babel.js --out-file jsonata-es5.js src/jsonata.js",
     "browserify": "node ./node_modules/browserify/bin/cmd.js jsonata-es5.js --outfile jsonata-es5.js --standalone jsonata",
     "minify-es5": "uglifyjs jsonata-es5.js -o jsonata-es5.min.js --compress --mangle"
   },
@@ -49,6 +50,7 @@
     "jsdoc": "^3.4.0",
     "mocha": "^5.0.0",
     "mocha-lcov-reporter": "^1.2.0",
+    "ncp": "^2.0.0",
     "request": "^2.81.0",
     "uglify-es": "^3.0.20"
   },
